+++ RELEASE-NOTES.md
@@ -19,3 +19,4 @@
 - forced Dark Mode to always on
 - updates to UI nav bar
 - fixed bug with harvest, stake, unstake and redeem buttons
+- add introduction text to homepage